+++ package.json
@@ -12,6 +12,8 @@
 		"test": "vitest",
 		"lint": "eslint \"**/*.{vue,ts}\"",
 		"lint:fix": "eslint \"**/*.{vue,ts}\" --fix",
+		"lint:deadcode": "knip --exclude binaries,dependencies,unlisted",
+		"lint:circular": "dpdm --exit-code circular:1 --no-tree --no-warning --progress false --transform './src/main.ts'",
 		"prettier": "prettier . --check",
 		"prettier:write": "prettier . --write",
 		"stylelint": "stylelint \"**/*.{vue,css}\"",
@@ -41,11 +43,13 @@
 		"@typescript-eslint/parser": "^8.15.0",
 		"@vitejs/plugin-vue": "^5.2.0",
 		"@vue/eslint-config-typescript": "^14.1.3",
+		"dpdm": "^3.14.0",
 		"eslint": "^9.15.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-prettier": "^5.2.1",
 		"eslint-plugin-vue": "^9.31.0",
 		"globals": "^15.12.0",
+		"knip": "^5.45.0",
 		"openapi-typescript": "^7.4.3",
 		"postcss": "^8.4.49",
 		"postcss-html": "^1.7.0",
@@ -55,6 +59,7 @@
 		"stylelint-config-recommended-vue": "^1.5.0",
 		"stylelint-config-standard": "^36.0.1",
 		"typescript": "5.6.3",
+		"typescript-eslint": "^8.26.0",
 		"vite": "^5.4.11",
 		"vite-plugin-eslint": "^1.8.1",
 		"vite-plugin-pwa": "^0.21.0",
