+++ src/api/controllers/auth/decorators.ts
@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import apiClient from '@/api/';
+import { ApiError } from '@/api';
+import { apiClient } from '../../client';
 import { ToastType } from '@/models';
 import router from '@/router';
 import { useProfileStore } from '@/store/profile';
 import { useToastStore } from '@/store/toast';
-import { apiError } from '@/utils/errorHandler';
 
 export type Func<R = any, FuncArgs extends any[] = any[]> = (...args: FuncArgs) => R;
 type Decorator<F extends Func = Func, DecoratorArgs extends any[] = any[]> = Func<
@@ -44,7 +44,7 @@ export function showErrorToast<F extends Func>(
 			const response = await method(...args);
 			return response;
 		} catch (err) {
-			const error = err as apiError;
+			const error = err as ApiError;
 			if (error) {
 				toastStore.push({
 					title: error.ru ?? error.message,
