+++ src/api/types.ts
@@ -1,0 +1,11 @@
+export type ApiError = {
+	status: string;
+	message: string;
+	ru: string;
+};
+
+export interface ErrorInfo {
+	url: string;
+	status: number;
+	message: string;
+}
